requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-limiter==2.0.0
pyjwt==2.1.0
python-dotenv==0.19.0
werkzeug==2.0.1
